// frontend/src/backend.d.ts
export type Time = bigint;
// ...
export interface ShopProduct {
    id: string;
    title: string;
    featured: boolean;
    inventory: bigint;
    published: boolean;
    popular: boolean;
    description: string;
    timestamp: Time;
    category: string;
    affiliateLink: string;
    price: number;
    images: Array<string>;
}